backend/pyproject.toml
[project]
name = "fingerprint-backend"
version = "1.0.0"
description = "Fingerprint — AI Learning Companion backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "google-cloud-firestore>=2.24.0",
    "google-genai>=1.66.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
